move utils to own folder
