+++ src/utils/emojis.ts
